pyproject.toml
[project]
name = "yoyopy"
version = "0.1.0"
description = "YoyoPod Connect - Screen-minimal streaming device for kids aged 6-12"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "YoyoPod Team" }
]
keywords = ["yoyopod", "yoyopy", "raspberry-pi", "iot", "audio-streaming", "kids-device"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pillow>=10.0.0",
    "pygame-ce>=2.4.0",
    "evdev>=1.6.1",
    "requests>=2.31.0",
    "paho-mqtt>=2.0.0",
    "loguru>=0.7.0",
    "displayhatmini>=0.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
yoyopy = "yoyopy.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["yoyopy"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
